v1.0.2
This commit adds a new build workflow to the project, updates the bug report template, adds a new lint workflow, and updates the .gitignore file. The go.yml file was removed from the workflows directory while coverage.txt was added. Additionally, some files were moved to the src directory and renamed. Finally, a TODO.md file was added to track future work.
