--- .github/workflows/go.yml
@@ -1,39 +1,0 @@
-# This workflow will build a golang project
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-go
-
-name: Go
-
-on:
-  push:
-    branches: ["master"]
-  pull_request:
-    branches: ["master"]
-
-jobs:
-  build:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-
-      - name: Set up Go
-        uses: actions/setup-go@v3
-        with:
-          go-version: 1.19
-
-      - name: Code Checks
-        run: go vet .
-
-      - name: Static Checks
-        uses: dominikh/staticcheck-action@v1.2.0
-        with:
-          version: "2022.1.1"
-          run: staticcheck .
-
-      - name: Build
-        run: go build plugin
-
-      - name: Test
-        run: go test plugin
-
-      - name: Upload coverage reports to Codecov
-        uses: codecov/codecov-action@v3
