+++ .gitignore
@@ -7,7 +7,7 @@
 *.dll
 *.so
 *.dylib
-*.txt
+# *.txt
 
 # Test binary, built with `go test -c`
 *.test
@@ -20,6 +20,13 @@
 
 # Go workspace file
 go.work
+*.code-workspace
 
 # Configs
 .vscode
+
+.env*
+.flaskenv*
+!.env.project
+!.env.vault
+/.env.vault
