+++ pull_request_template.md
@@ -1,3 +1,31 @@
-# Pull Request Template
+# Pull Request
 
-## Coming Soon
+## Description
+
+Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change.
+
+## Pull request type
+
+Please check the change your PR introduces:
+
+- [ ] Bugfix
+- [ ] Feature
+- [ ] Code style update (formatting, renaming)
+- [ ] Refactoring (no functional changes, no API changes)
+- [ ] Build related changes
+- [ ] Documentation content changes
+- [ ] Other (please describe):
+
+## What is the current behavior?
+
+-
+
+Issue Number: N/A
+
+## What is the new behavior?
+
+-
+
+## Other information
+
+Anything else to know?
